requirements.txt
mcp[cli]
mcpo
pip-system-certs
python-dotenv